requirements.txt
Django==1.5
Fabric==1.5.3
Jinja2==2.6
Pygments==1.6
Unidecode==0.04.12
argparse==1.2.1
blinker==1.2
docutils==0.10
feedgenerator==1.5
paramiko==1.9.0
pelican==3.1.1
pycrypto==2.6
pytz==2012j
six==1.2.0
unicode-slugify==0.1
wsgiref==0.1.2
